// libs/user/src/lib/user/types/frnds_app_action.types.ts
export enum FrndsAppInitActionTypes {
  FRNDS_INIT = '[FRNDS APP] FRNDS APP INIT',
  FRNDS_INIT_SUCCESS = '[FRNDS APP] FRNDS INIT SUCCESS',
  FRNDS_INIT_ERROR = '[FRNDS APP] FRNDS INIT ERROR',
}

export enum FrndsAppSelectAllUserActionTypes {
  FRNDS_GET_ALL_CLICK = '[FRNDS APP] FRNDS GET_ALL CLICK',
  FRNDS_GET_ALL_INIT = '[FRNDS APP] FRNDS GET_ALL INIT',
  FRNDS_GET_ALL_SUCCESS = '[FRNDS APP] FRNDS GET_ALL SUCCESS',
  FRNDS_GET_ALL_ERROR = '[FRNDS APP] FRNDS GET_ALL ERROR',
}

export enum FrndsAppSelectUserByQueryActionTypes {
  FRNDS_GET_BY_QUERY_CLICK = '[FRNDS APP] FRNDS GET_BY_QUERY CLICK',
  FRNDS_GET_BY_QUERY_INIT = '[FRNDS APP] FRNDS GET_BY_QUERY INIT',
  FRNDS_GET_BY_QUERY_SUCCESS = '[FRNDS APP] FRNDS GET_BY_QUERY SUCCESS',
  FRNDS_GET_BY_QUERY_ERROR = '[FRNDS APP] FRNDS GET_BY_QUERY ERROR',
}

export enum FrndsAppSelectUserByIdActionTypes {
  FRNDS_GET_BY_ID_CLICK = '[FRNDS APP] FRNDS GET_BY_ID CLICK',
  FRNDS_GET_BY_ID_INIT = '[FRNDS APP] FRNDS GET_BY_ID INIT',
  FRNDS_GET_BY_ID_SUCCESS = '[FRNDS APP] FRNDS GET_BY_ID SUCCESS',
  FRNDS_GET_BY_ID_ERROR = '[FRNDS APP] FRNDS GET_BY_ID ERROR',
}

export enum FrndsAppNewUserActionTypes {
  FRNDS_NEW_USER_CLICK = '[FRNDS APP] NEW_USER_CLICK',
  FRNDS_NEW_USER_INIT = '[FRNDS APP] NEW_USER INIT',
  FRNDS_NEW_USER_SUCCESS = '[FRNDS APP] NEW_USER SUCCESS',
  FRNDS_NEW_USER_ERROR = '[FRNDS APP] NEW_USER ERROR',
}

export enum FrndsAppEditUserActionTypes {
  FRNDS_EDIT_USER_CLICK = '[FRNDS APP] EDIT USER_CLICK',
  FRNDS_EDIT_USER_INIT = '[FRNDS APP] EDIT USER INIT',
  FRNDS_EDIT_USER_SUCCESS = '[FRNDS APP] EDIT USER SUCCESS',
  FRNDS_EDIT_USER_ERROR = '[FRNDS APP] EDIT USER ERROR'
}

export enum FrndsAppDeleteUserActionTypes {
  FRNDS_DELETE_USER_CLICK = '[FRNDS APP] DELETE USER_CLICK',
  FRNDS_DELETE_USER_INIT = '[FRNDS APP] DELETE USER INIT',
  FRNDS_DELETE_USER_SUCCESS = '[FRNDS APP] DELETE USER SUCCESS',
  FRNDS_DELETE_USER_ERROR = '[FRNDS APP] DELETE USER ERROR'
}

export enum FrndsAppChartGenerateActionTypes {
  FRNDS_CHARTS_GEN_INIT = '[FRNDS APP] CHARTS INIT',
  FRNDS_CHARTS_GEN_SUCCESS = '[FRNDS APP] CHARTS SUCCESS',
  FRNDS_CHARTS_GEN_FAILURE = '[FRNDS APP] CHARTS FAILURE'
}

export enum FrndsAppClearUserActionTypes {
  FRNDS_CLEAR_USER_CLICK = '[FRNDS APP] CLEAR USER_CLICK',
}